Simplify stateless_mode! API
Instead of passing wheel leds and a side, a slice is supplied of the
relevant LEDs, so most of the time, side is not needed.

Consequently side is removed from the render call.

The only pattern this caused a problem for was the fuck boris one,
but that was easily re-implemented using the other bitmap text
mode support that I've subsequently implemented.