Rearrange most imports so this is mostly a library crate. Add benchmark.
This started as benchmarking work, but in order to access the function
I wanted from benchmarking, I needed to rearrange the crate structure to
be be mostly a library, used by main.rs and the benchmark.

The benchmark was to examine the likely effects on rendering speed if
I can run SPI at faster than 1MHz (by changing the level shifting
hardware).

At 1MHz, a dump of all pixels to SPI takes 1.7ms. At 4MHz, it takes 500us,
and at 32MHz it takes 187us. This looks very hopeful.